+++ src/app/core/services-firebase/index.ts
@@ -1,4 +1,5 @@
 export * from './mail.service';
+export * from './official-outage.service';
 export * from './report.service';
 export * from './base.service';
 export * from './auth.service';